plugin support for persistent volumes in group
