+++ pkg/plugin/costing.go
@@ -20,7 +20,6 @@ package plugin
 import (
 	"fmt"
 
-	"github.com/vmware/purser/pkg/plugin/crd"
 	"github.com/vmware/purser/pkg/plugin/metrics"
 
 	"k8s.io/api/core/v1"
@@ -229,105 +228,4 @@ func getPodsCost(pods []*Pod) []*Pod {
 	pvcs = collectPersistentVolumeClaims(pvcs)
 	pods = calculateCost(pods, pvcs)
 	return pods
-}
-
-// GetGroupDetails returns aggregated metrics (cpu, memory, storage) and cost (total, cpu, memory and storage) of a Group
-func GetGroupDetails(group *crd.Group) (*metrics.GroupMetrics, *metrics.GroupMetrics, *Cost) {
-	// TODO: include storage in group details
-	price := GetUserCosts()
-
-	currentTime := getCurrentTime()
-	monthStartTime := getCurrentMonthStartTime()
-
-	podsDetails := group.Spec.PodsDetails
-	var totalCPURequest, totalCPULimit, totalMemoryRequest, totalMemoryLimit, totalStorageClaimed float64
-	// [PIT] Point In Time metrics for the group
-	var pitCPURequest, pitCPULimit, pitMemoryRequest, pitMemoryLimit, pitStorageClaimed float64
-	for _, podDetails := range podsDetails {
-		startTime := podDetails.StartTime
-		endTime := podDetails.EndTime
-
-		podActiveHours := currentMonthActiveTimeInHoursMulti(startTime, endTime, currentTime, monthStartTime)
-
-		podMetrics := CalculatePodMetricsFromPodDetails(podDetails)
-
-		podCPURequest := float64(podMetrics.CPURequest.Value())
-		podMemRequest := bytesToGB(podMetrics.MemoryRequest.Value())
-		podCPULimit := float64(podMetrics.CPULimit.Value())
-		podMemLimit := bytesToGB(podMetrics.MemoryLimit.Value())
-
-		// Pod is alive
-		if endTime.IsZero() {
-			pitCPULimit += podCPULimit
-			pitCPURequest += podCPURequest
-			pitMemoryLimit += podMemLimit
-			pitMemoryRequest += podMemRequest
-		}
-
-		totalCPURequest += podCPURequest * podActiveHours
-		totalMemoryRequest += podMemRequest * podActiveHours
-		totalCPULimit += podCPULimit * podActiveHours
-		totalMemoryLimit += podMemLimit * podActiveHours
-
-		// TODO: find podStorage
-		podStorageClaimed := 0.0
-		totalStorageClaimed += podStorageClaimed * podActiveHours
-	}
-
-	totalCPUCost := price.CPU * totalCPURequest
-	totalMemoryCost := price.Memory * totalMemoryRequest
-	totalStorageCost := price.Storage * totalStorageClaimed
-
-	totalCumulativeCost := totalCPUCost + totalMemoryCost + totalStorageCost
-
-	mtdGroupMetrics := &metrics.GroupMetrics{
-		CPULimit:       totalCPULimit,
-		MemoryLimit:    totalMemoryLimit,
-		CPURequest:     totalCPURequest,
-		MemoryRequest:  totalMemoryRequest,
-		StorageClaimed: totalStorageClaimed,
-	}
-
-	pitGroupMetrics := &metrics.GroupMetrics{
-		CPULimit:       pitCPULimit,
-		MemoryLimit:    pitMemoryLimit,
-		CPURequest:     pitCPURequest,
-		MemoryRequest:  pitMemoryRequest,
-		StorageClaimed: pitStorageClaimed,
-	}
-
-	cost := &Cost{
-		TotalCost:   totalCumulativeCost,
-		CPUCost:     totalCPUCost,
-		MemoryCost:  totalMemoryCost,
-		StorageCost: totalStorageCost,
-	}
-
-	return pitGroupMetrics, mtdGroupMetrics, cost
-}
-
-// CalculatePodMetricsFromPodDetails returns PodMetrics given its details(container metrics)
-func CalculatePodMetricsFromPodDetails(podDetails *crd.PodDetails) *metrics.Metrics {
-	CPURequest := resource.Quantity{}
-	MemoryRequest := resource.Quantity{}
-	CPULimit := resource.Quantity{}
-	MemoryLimit := resource.Quantity{}
-	for _, container := range podDetails.Containers {
-		addResourceAToResourceB(container.Metrics.CPURequest, &CPURequest)
-		addResourceAToResourceB(container.Metrics.MemoryRequest, &MemoryRequest)
-		addResourceAToResourceB(container.Metrics.MemoryLimit, &MemoryLimit)
-		addResourceAToResourceB(container.Metrics.CPULimit, &CPULimit)
-	}
-	return &metrics.Metrics{
-		CPULimit:      &CPULimit,
-		MemoryLimit:   &MemoryLimit,
-		CPURequest:    &CPURequest,
-		MemoryRequest: &MemoryRequest,
-	}
-}
-
-func addResourceAToResourceB(resA, resB *resource.Quantity) {
-	if resA != nil {
-		resB.Add(*resA)
-	}
 }
